Updated BE to work with module intead.
