+++ BackEnd/package.json
@@ -4,6 +4,7 @@
   "main": "index.js",
   "author": "Paulo Braz",
   "license": "MIT",
+  "type": "module",
   "dependencies": {
     "body-parser": "^1.20.2",
     "express": "^4.18.2",
