+++ BackEnd/server.js
@@ -1,6 +1,6 @@
-const express = require('express');
-const bodyParser = require('body-parser');
-const mongoDb = require('./mongoose');
+import express from 'express';
+import bodyParser from 'body-parser';
+import { addNewUserToList, getUsersList } from "./mongooseDb.js";
 
 const app = express();
 
@@ -11,10 +11,15 @@ app.use((req, res, next) => {
     res.setHeader('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Authorization');
     res.setHeader('Access-Control-Allow-Methods', 'GET, POST, PATCH, DELETE');
     next();
+});
+
+app.get('/', (req, res, next) => {
+    console.log('It works!');
+    res.json('It works');
+    next();
 })
 
-app.post('/userlist', mongoDb.addNewUserToList);
-app.get('/userlist', mongoDb.getUsersList);
-
+app.post('/userlist', addNewUserToList);
+app.get('/userlist', getUsersList);
 
 app.listen(5000);
